% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.R
\name{c2d}
\alias{c2d}
\title{convert model from continuous to discrete time}
\usage{
c2d(A, B, Ts)
}
\arguments{
\item{A}{State matrix}

\item{B}{Input matrix}

\item{Ts}{Timestep}
}
\value{
A discreted state-space model
}
\description{
Takes in state-space model and discretizes it
}
\examples{
A <- matrix(c(0.51,0.22,0.47,0.78),nrow = 2, byrow = T)
B <- matrix(c(0.61,0.83,0.25,0.39),nrow = 2, byrow = T)
csys <- d2c(A,B,600)
c2d(csys$A,csys$B,600)
c2d(csys$A,csys$B,1200)
}
