% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_consumption_units.R
\docType{data}
\name{consumption}
\alias{consumption}
\title{Consumption Unit}
\format{An object of R6Class}
\usage{
model <- consumption$new()

model$setvalue(ssmodel,...,x0,name = NA)

model$ifcon() #if the state-space is continuous or not
}
\description{
Build, set value of the Consumption Unit(i.e. building )
}
\examples{
search `EMPC` on github, find more examples
}
\keyword{datasets}
