% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_ehubmpc.R
\docType{data}
\name{ehubmpc}
\alias{ehubmpc}
\title{ehubmpc controller}
\format{list of equipment power(U),temperature values(Y),battery charge/discharge power(Ub),
net electricity power(E), battery capacity(Xb)}
\usage{
model <- ehubmpc$new()
model$initialize()
model$set_parameters(N,Tsim,obj,cost,ymin,ymax,yref,ECR,umax,umin)
model$print_para()
model$set_mpc_constraint()
model$solve_mpc(unit.index = c(1,0),control = ecos.control())
}
\arguments{
\item{N}{prediction horizon}

\item{Tsim}{simulation horizon; Tsim >= N}

\item{obj}{control purpose; obj should be one of "cost" or "comfort". "cost" means "cost-optimal"
econimic mpc while "comfort" means temperature/humidity control. Different value respresents
different objective in linear programming}

\item{cost}{electricity price; it should be a matrix of electricity price whether the price is changed
or not}

\item{ymin/ymax}{lower/upper limit of temperature/humidity; a matrix}

\item{yref}{reference y, i.e. temperature setpoint; a matrix}

\item{ECR}{penalty factor for slack variable s; default 1e6}

\item{umax/umin}{lower/upper limits of heating/cooling power}

\item{unit.index}{a vector indicates whether storage or production units exist or not.
Default value is c(1,0), which means there is storage unit without production unit.}

\item{control}{ecos.control; control parameters for ECOS solver. More details can be seen in ECOSolver
package}
}
\description{
solve ehubmpc for "cost-optimal" control in energyhub with storage(i.e. battery, ice storage)
or/and production(i.e. PV, wind turbine) units
}
\note{
Two battery kinds, namely simple model and lead-acid battery model are support at present.
}
\examples{
search `EMPC` on github, find more examples
}
\keyword{datasets}
