% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_mpc.R
\docType{data}
\name{mpc}
\alias{mpc}
\title{mpc controller}
\format{list of equipment power(U) and temperature values(Y)}
\usage{
model <- mpc$new()
model$initialize()
model$set_parameters(N,Tsim,obj,cost,ymin,ymax,yref,ECR,umax,umin)
model$print_para()
model$set_mpc_constraint()
model$solve_mpc(control = ecos.control())
}
\arguments{
\item{N}{prediction horizon}

\item{Tsim}{simulation horizon; Tsim >= N}

\item{obj}{control purpose; obj should be one of "cost" or "comfort". "cost" means "cost-optimal"
econimic mpc while "comfort" means temperature/humidity control. Different value respresents
different objective in linear programming}

\item{cost}{electricity price; it should be a matrix of electricity price whether the price is changed
or not}

\item{ymin/ymax}{lower/upper limit of temperature/humidity; a matrix}

\item{yref}{reference y, i.e. temperature setpoint; a matrix}

\item{ECR}{penalty factor for slack variable s; default 1e6}

\item{umax/umin}{lower/upper limits of heating/cooling power}

\item{control}{ecos.control; control parameters for ECOS solver. More details can be seen in ECOSolver
package}
}
\description{
solve mpc for building climate control or "cost-optimal" control
}
\note{
If your focus is energyhub which contains building unit, production or storage unit
, ehubmpc object maybe a better choice. mpc only supports building unit without other units involved.
Besides, time-variant constraints, such as temperature setpoints and range, electricity cost,
equipment power (heating/cooling power) are supported.
}
\examples{
search `EMPC` on github, find more examples
}
\keyword{datasets}
