% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_production_units.R
\docType{data}
\name{production}
\alias{production}
\title{Production Unit}
\format{An object of R6Class}
\usage{
model <- production$new()

model$setvalue(generation,timestep)
}
\description{
Build, set value of the Production Unit(i.e. PV or WindTurbine)
}
\examples{
search `EMPC` on github, find more examples
}
\keyword{datasets}
