% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_storage_units.R
\docType{data}
\name{storage}
\alias{storage}
\title{Storage Unit}
\format{An object of R6Class}
\usage{
model <- storage$new()

model$setvalue(ssmodel,...,A_cst,B_cst,h_cst)

model$ifcon() #if the state-space is continuous or not
}
\description{
Build, set value of the Storage Unit(i.e. building )
}
\examples{
search `EMPC` on github, find more examples
}
\keyword{datasets}
