% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_production_units.R
\docType{data}
\name{production}
\alias{production}
\title{Production Unit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
production
}
\description{
Build, set value of the Production Unit(i.e. PV or WindTurbine)
}
\examples{
#model <- production$new()

#model$setvalue(generation,timestep)

}
\keyword{datasets}
