% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_storage_units.R
\docType{data}
\name{storage}
\alias{storage}
\title{Storage Unit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
storage
}
\description{
Build, set value of the Storage Unit(i.e. building )
}
\examples{

#model <- storage$new()

#model$setvalue(ssmodel,...,A_cst,B_cst,h_cst)

#model$ifcon() #if the state-space is continuous or not

}
\keyword{datasets}
